updated balances from floats to integers in response to the schema

+++ lib/account.rb
@@ -17,7 +17,7 @@ module Bank
 
     def self.all
       CSV.read("support/accounts.csv").collect do |account|
-        Account.new(id: account[0], balance: account[1].to_f, open_date: account[2])
+        Account.new(id: account[0], balance: account[1].to_i, open_date: account[2])
       end
     end
 
@@ -25,7 +25,7 @@ module Bank
       target_account = nil
       CSV.read("support/accounts.csv").each do |account|
         if account[0] == id
-          target_account = Account.new(id: account[0], balance: account[1].to_f, open_date: account[2])
+          target_account = Account.new(id: account[0], balance: account[1].to_i, open_date: account[2])
         end
       end
       raise ArgumentError.new("Invalid account ID.") if target_account.nil?
